Switch to Powerlevel10k zsh theme (#25)
Fixes #24

Having my own theme isn't maintainable, and it pales in comparison to some of the fancier themes out there on the internet. Back in the day, I tried using Powerlevel9k as a replacement theme, but it was too slow to be usable. Now that we have Powerlevel10k, it's a fantastic replacement for my current zsh theme.