Revise instructions to use prepatched fonts on mac (#28)
We're using the patching strategy instead of the fallback strategy for Mac. That way we don't need to disable csrutils and go through a complicated setup process. I still need to manually verify that this works on a new machine, but I'm relatively confident there's not much else to this process.